Clean up and improve cgcore.html
This cleans up the HTML and CSS to be more HTML5 compliant and use
Bootstrap helper CSS classes instead of hardcoding margins, etc. I
also made minor improvements to the style based on the Dublin Core
website layout and added a visual indicator for the left navbar.

I have not changed any of the language, grammar, or schema in this.